Backport "--disable-xmldoc" fix  (ASTERISK-26639)
This is a patch contributed to upstream that is not released yet.  As
we treat all Asterisk compilations as though they were on an embedded
platform in order to conserve and optimize infrastructure resources, we
do not ship Asterisk documentation and therefore Asterisk should not
fail to run if documentation is not present.  However, as of 07/18/17,
—disable-xmldic is broken and does not work, which this patch resolves.

From Corey Farrell who contributed the patch to Digium:

“The configure option to disable XML documentation does not currently
work.  This patch makes it effective, but also causes an ABI change by
removing the ast_xmldoc_* symbols.”

https://issues.asterisk.org/jira/secure/attachment/55281/0001-Build-Syst
em-Fix-disable-xmldoc-option.patch